Move over to lilybird (#31)
This pr changes discord.js in favor of lilybird.

### Notes:

This pr also moves `MESSAGE_PREFIX` and `BUN_ONLY_CHANNEL_ID` to
`process.env` instead of being constants.

`allowed_mentions` are not supported yet, i will look over them
eventually but for now it should work fine anyways

This pr should not be merged yet, lilybird 0.2.3 doesnt not have
`client.ping` this will come in 0.2.4
