+++ src/index.ts
@@ -1,11 +1,30 @@
-import "./loaders/tags.ts";
+import "./loaders/tags";
 import "./commands";
 import "./listeners";
 
-import { Bubu } from "./structs/Client.ts";
+import { createHandler } from "@lilybird/handlers";
+import { createClient, Intents } from "lilybird";
 
 // Make sure bubu will not crash
 process.on("unhandledRejection", console.error);
 process.on("uncaughtException", console.error);
 
-Bubu.login(process.env.DISCORD_BOT_TOKEN).catch(console.error);
+const listeners = await createHandler({
+    prefix: process.env.MESSAGE_PREFIX,
+    dirs: {
+        messageCommands: `${import.meta.dir}/message-commands`,
+        slashCommands: `${import.meta.dir}/commands`,
+        listeners: `${import.meta.dir}/listeners`
+    }
+});
+
+await createClient({
+    token: process.env.DISCORD_BOT_TOKEN,
+    intents: [
+        Intents.GUILDS,
+        Intents.GUILD_MESSAGES,
+        Intents.MESSAGE_CONTENT,
+        Intents.GUILD_MEMBERS
+    ],
+    ...listeners
+});
