--- src/listeners/index.ts
@@ -1,5 +1,0 @@
-import "./ready.ts";
-import "./interaction_create.ts";
-import "./message_create.ts";
-import "./message_update.ts";
-import "./nickname_moderation.ts";
