+++ src/listeners/member_add.ts
@@ -1,0 +1,9 @@
+import { moderateNick } from "../util.ts";
+import { Event } from "@lilybird/handlers";
+
+export default {
+    event: "guildMemberAdd",
+    run: (member) => {
+        moderateNick(member);
+    },
+} satisfies Event<"guildMemberAdd">
