+++ src/listeners/member_update.ts
@@ -1,0 +1,9 @@
+import { moderateNick } from "../util.ts";
+import { Event } from "@lilybird/handlers";
+
+export default {
+    event: "guildMemberUpdate",
+    run: (member) => {
+        moderateNick(member);
+    },
+} satisfies Event<"guildMemberUpdate">
