--- src/structs/Listener.ts
@@ -1,10 +1,0 @@
-import type { ClientEvents } from "discord.js";
-
-export interface Listener<E extends keyof ClientEvents> {
-  event: E;
-  once?: boolean;
-  run: (
-    ...args: ClientEvents[E]
-  ) => any;
-}
-
