+++ src/util.ts
@@ -1,10 +1,23 @@
-export function safeSlice<T>(input: T, length: number) {
-  // @ts-expect-error i know where im using it
-  return input.length > length ? input.slice(0, length) : input;
+import { GuildMember } from "lilybird";
+
+export function safeSlice<T extends string | Array<any>>(input: T, length: number): T {
+    return <T>(input.length > length ? input.slice(0, length) : input);
 }
 
 export async function silently<T>(value: Promise<T>) {
-  try {
-    await value;
-  } catch {}
+    try {
+        await value;
+    } catch { }
+}
+
+export async function moderateNick(member: GuildMember) {
+    let name = member.nick ?? member.user.username;
+    const normalizedName = name.normalize("NFKC").replace(/^[!$#@%^`&*()]+/, "");
+
+    if (name === normalizedName) return;
+
+    silently(member.modify({
+        nick: normalizedName,
+        reason: "lame username"
+    }));
 }
